refactor: LazyTest
