+++ test/LazyTest/Utils/TestException.cs
@@ -1,0 +1,16 @@
+namespace LazyTest.Utils;
+
+using System;
+
+/// <summary>
+/// Test exception class.
+/// </summary>
+public class TestException : Exception
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TestException"/> class.
+    /// </summary>
+    public TestException()
+    {
+    }
+}
